all reqs complete

+++ controllers/clothingItems.js
@@ -4,6 +4,7 @@ const {
   handleCastError,
   handleNotFoundError,
   handleGenericError,
+  handleForbiddenError,
 } = require("../utils/errors");
 
 const getItems = (req, res) => {
@@ -28,14 +29,26 @@ const createItem = (req, res) => {
 
 const deleteItem = (req, res) => {
   const { itemId } = req.params;
+  const userId = req.user._id;
 
-  ClothingItem.findByIdAndDelete(itemId)
+  ClothingItem.findById(itemId)
     .orFail()
-    .then((item) => res.status(200).send(item))
+    .then((item) => {
+      if (String(item.owner) !== String(userId)) {
+        const err = new Error("You are not allowed to delete this item");
+        err.name = "Forbidden";
+        throw err;
+      }
+
+      return ClothingItem.findByIdAndDelete(itemId);
+    })
+    .then((deletedItem) => res.status(200).send(deletedItem))
     .catch((err) => {
       if (err.name === "DocumentNotFoundError")
         return handleNotFoundError(err, res);
       if (err.name === "CastError") return handleCastError(err, res);
+      if (err.name === "Forbidden") return handleForbiddenError(err, res);
+
       return handleGenericError(err, res);
     });
 };
